+++ eterra/services/dnp3_service/dnp3/requirements.txt.txt
@@ -1,0 +1,4 @@
+pydnp3==0.1.0
+PyYAML==3.12
+pytz==2016.6.1
+gridappsd-python
