--- eterra/services/requirements.txt
@@ -1,4 +1,0 @@
-
-pydnp3==0.1.0
-PyYAML==3.12
-
